tests: Factor into separate files
Factors the monolithic `tests/lib.rs` into separate files, in hopes of
making the code more maintainable.
